requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
supabase